+++ demo/vanilla/app.js
@@ -85,6 +85,7 @@ app.use("/path-curves", express.static("./node_modules/tsparticles-path-curves")
 app.use("/path-polygon", express.static("./node_modules/tsparticles-path-polygon"));
 app.use("/path-perlin-noise", express.static("./node_modules/tsparticles-path-perlin-noise"));
 app.use("/path-simplex-noise", express.static("./node_modules/tsparticles-path-simplex-noise"));
+app.use("/plugin-hsv-color", express.static("./node_modules/tsparticles-plugin-hsv-color"));
 app.use("/plugin-infection", express.static("./node_modules/tsparticles-plugin-infection"));
 app.use("/preset-big-circles", express.static("./node_modules/tsparticles-preset-big-circles"));
 app.use("/preset-bubbles", express.static("./node_modules/tsparticles-preset-bubbles"));