+++ editor/webpack.config.js
@@ -18,6 +18,7 @@ const getJsConfig = (entry) => {
 
   return {
     entry: entry,
+    mode: "production",
     output: {
       path: path.resolve(__dirname, "dist"),
       filename: "[name].js",