+++ engine/src/Utils/ColorUtils.ts
@@ -1,28 +1,21 @@
-import type {
-    IColor,
-    IHsl,
-    IHsla,
-    IHsv,
-    IHsva,
-    IRangeColor,
-    IRangeHsl,
-    IRangeHsv,
-    IRangeRgb,
-    IRangeValueColor,
-    IRgb,
-    IRgba,
-    IValueColor,
-} from "../Core/Interfaces/Colors";
+import type { IColor, IHsl, IHsla, IRangeColor, IRgb, IRgba } from "../Core/Interfaces/Colors";
 import { getRangeValue, mix, randomInRange, setRangeValue, tspRandom } from "./NumberUtils";
 import { midColorValue, randomColorValue } from "../Core/Utils/Constants";
 import { AnimationStatus } from "../Enums/AnimationStatus";
 import type { HslAnimation } from "../Options/Classes/HslAnimation";
 import type { IColorAnimation } from "../Options/Interfaces/IColorAnimation";
+import type { IColorManager } from "../Core/Interfaces/IColorManager";
 import type { IOptionsColor } from "../Options/Interfaces/IOptionsColor";
 import type { IParticle } from "../Core/Interfaces/IParticle";
 import type { IParticleHslAnimation } from "../Core/Interfaces/IParticleHslAnimation";
 import type { IParticleValueAnimation } from "../Core/Interfaces/IParticleValueAnimation";
 import { itemFromArray } from "./Utils";
+
+const colorManagers = new Map<string, IColorManager>();
+
+export function addColorManager(key: string, manager: IColorManager): void {
+    colorManagers.set(key, manager);
+}
 
 /**
  * Converts hue to RGB values.
@@ -60,59 +53,27 @@ function hue2rgb(p: number, q: number, t: number): number {
  * @param input A string that represents a color.
  */
 function stringToRgba(input: string): IRgba | undefined {
-    if (input.startsWith("rgb")) {
-        const regex = /rgba?\(\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*(,\s*([\d.]+)\s*)?\)/i,
-            result = regex.exec(input);
-
-        return result
-            ? {
-                  a: result.length > 4 ? parseFloat(result[5]) : 1,
-                  b: parseInt(result[3], 10),
-                  g: parseInt(result[2], 10),
-                  r: parseInt(result[1], 10),
-              }
-            : undefined;
-    } else if (input.startsWith("hsl")) {
-        const regex = /hsla?\(\s*(\d+)\s*,\s*(\d+)%\s*,\s*(\d+)%\s*(,\s*([\d.]+)\s*)?\)/i,
-            result = regex.exec(input);
-
-        return result
-            ? hslaToRgba({
-                  a: result.length > 4 ? parseFloat(result[5]) : 1,
-                  h: parseInt(result[1], 10),
-                  l: parseInt(result[3], 10),
-                  s: parseInt(result[2], 10),
-              })
-            : undefined;
-    } else if (input.startsWith("hsv")) {
-        const regex = /hsva?\(\s*(\d+)°\s*,\s*(\d+)%\s*,\s*(\d+)%\s*(,\s*([\d.]+)\s*)?\)/i,
-            result = regex.exec(input);
-
-        return result
-            ? hsvaToRgba({
-                  a: result.length > 4 ? parseFloat(result[5]) : 1,
-                  h: parseInt(result[1], 10),
-                  s: parseInt(result[2], 10),
-                  v: parseInt(result[3], 10),
-              })
-            : undefined;
-    } else {
-        const shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])([a-f\d])?$/i,
-            hexFixed = input.replace(shorthandRegex, (_, r, g, b, a) => {
-                return r + r + g + g + b + b + (a !== undefined ? a + a : "");
-            }),
-            regex = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})?$/i,
-            result = regex.exec(hexFixed);
-
-        return result
-            ? {
-                  a: result[4] !== undefined ? parseInt(result[4], 16) / 0xff : 1,
-                  b: parseInt(result[3], 16),
-                  g: parseInt(result[2], 16),
-                  r: parseInt(result[1], 16),
-              }
-            : undefined;
+    for (const [, manager] of colorManagers) {
+        if (input.startsWith(manager.stringPrefix)) {
+            return manager.parseString(input);
+        }
     }
+
+    const shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])([a-f\d])?$/i,
+        hexFixed = input.replace(shorthandRegex, (_, r, g, b, a) => {
+            return r + r + g + g + b + b + (a !== undefined ? a + a : "");
+        }),
+        regex = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})?$/i,
+        result = regex.exec(hexFixed);
+
+    return result
+        ? {
+              a: result[4] !== undefined ? parseInt(result[4], 16) / 0xff : 1,
+              b: parseInt(result[3], 16),
+              g: parseInt(result[2], 16),
+              r: parseInt(result[1], 16),
+          }
+        : undefined;
 }
 
 /**
@@ -138,35 +99,12 @@ export function rangeColorToRgb(input?: string | IRangeColor, index?: number, us
         });
     }
 
-    const colorValue = color.value as IRangeValueColor,
-        rgbColor = colorValue.rgb ?? (color.value as IRangeRgb);
+    for (const [, manager] of colorManagers) {
+        const res = manager.handleRangeColor(color);
 
-    if (rgbColor.r !== undefined) {
-        return {
-            r: getRangeValue(rgbColor.r),
-            g: getRangeValue(rgbColor.g),
-            b: getRangeValue(rgbColor.b),
-        };
-    }
-
-    const hslColor = colorValue.hsl ?? (color.value as IRangeHsl);
-
-    if (hslColor.h !== undefined && hslColor.l !== undefined) {
-        return hslToRgb({
-            h: getRangeValue(hslColor.h),
-            l: getRangeValue(hslColor.l),
-            s: getRangeValue(hslColor.s),
-        });
-    }
-
-    const hsvColor = colorValue.hsv ?? (color.value as IRangeHsv);
-
-    if (hsvColor.h !== undefined && hsvColor.v !== undefined) {
-        return hsvToRgb({
-            h: getRangeValue(hsvColor.h),
-            s: getRangeValue(hsvColor.s),
-            v: getRangeValue(hsvColor.v),
-        });
+        if (res) {
+            return res;
+        }
     }
 
     return;
@@ -195,26 +133,13 @@ export function colorToRgb(input?: string | IColor, index?: number, useIndex = t
         });
     }
 
-    const colorValue = color.value as IValueColor,
-        rgbColor = colorValue.rgb ?? (color.value as IRgb);
+    for (const [, manager] of colorManagers) {
+        const res = manager.handleColor(color);
 
-    if (rgbColor.r !== undefined) {
-        return rgbColor;
+        if (res) {
+            return res;
+        }
     }
-
-    const hslColor = colorValue.hsl ?? (color.value as IHsl);
-
-    if (hslColor.h !== undefined && hslColor.l !== undefined) {
-        return hslToRgb(hslColor);
-    }
-
-    const hsvColor = colorValue.hsv ?? (color.value as IHsv);
-
-    if (hsvColor.h !== undefined && hsvColor.v !== undefined) {
-        return hsvToRgb(hsvColor);
-    }
-
-    return;
 }
 
 /**
@@ -361,190 +286,6 @@ export function hslaToRgba(hsla: IHsla): IRgba {
 }
 
 /**
- * Converts a Hue Saturation Lightness ([[IHsl]]) object in a [[IHsv]] object
- * @param hsl the Hue Saturation Lightness ([[IHsl]]) object
- * @returns the [[IHsv]] object
- */
-export function hslToHsv(hsl: IHsl): IHsv {
-    const l = hsl.l / 100,
-        sl = hsl.s / 100,
-        v = l + sl * Math.min(l, 1 - l),
-        sv = !v ? 0 : 2 * (1 - l / v);
-
-    return {
-        h: hsl.h,
-        s: sv * 100,
-        v: v * 100,
-    };
-}
-
-/**
- * Converts a Hue Saturation Lightness ([[IHsla]]) object in a [[IHsva]] object
- * @param hsla the Hue Saturation Lightness ([[IHsla]]) object
- * @returns the [[IHsva]] object
- */
-export function hslaToHsva(hsla: IHsla): IHsva {
-    return {
-        a: hsla.a,
-        ...hslToHsv(hsla),
-    };
-}
-
-/**
- * Converts a Hue Saturation Lightness ([[IHsv]]) object in a [[IHsl]] object
- * @param hsv the Hue Saturation Lightness ([[IHsv]]) object
- * @returns the [[IHsl]] object
- */
-export function hsvToHsl(hsv: IHsv): IHsl {
-    const v = hsv.v / 100,
-        sv = hsv.s / 100,
-        l = v * (1 - sv / 2),
-        sl = l === 0 || l === 1 ? 0 : (v - l) / Math.min(l, 1 - l);
-
-    return {
-        h: hsv.h,
-        l: l * 100,
-        s: sl * 100,
-    };
-}
-
-/**
- * Converts a Hue Saturation Lightness ([[IHsva]]) object in a [[IHsla]] object
- * @param hsva the Hue Saturation Lightness ([[IHsva]]) object
- * @returns the [[IHsla]] object
- */
-export function hsvaToHsla(hsva: IHsva): IHsla {
-    return {
-        a: hsva.a,
-        ...hsvToHsl(hsva),
-    };
-}
-
-/**
- * Converts a Hue Saturation Lightness ([[IHsv]]) object in a [[IRgb]] object
- * @param hsv the Hue Saturation Lightness ([[IHsv]]) object
- * @returns the [[IRgb]] object
- */
-export function hsvToRgb(hsv: IHsv): IRgb {
-    const result: IRgb = { b: 0, g: 0, r: 0 },
-        hsvPercent = {
-            h: hsv.h / 60,
-            s: hsv.s / 100,
-            v: hsv.v / 100,
-        },
-        c = hsvPercent.v * hsvPercent.s,
-        x = c * (1 - Math.abs((hsvPercent.h % 2) - 1));
-
-    let tempRgb: IRgb | undefined;
-
-    if (hsvPercent.h >= 0 && hsvPercent.h <= 1) {
-        tempRgb = {
-            r: c,
-            g: x,
-            b: 0,
-        };
-    } else if (hsvPercent.h > 1 && hsvPercent.h <= 2) {
-        tempRgb = {
-            r: x,
-            g: c,
-            b: 0,
-        };
-    } else if (hsvPercent.h > 2 && hsvPercent.h <= 3) {
-        tempRgb = {
-            r: 0,
-            g: c,
-            b: x,
-        };
-    } else if (hsvPercent.h > 3 && hsvPercent.h <= 4) {
-        tempRgb = {
-            r: 0,
-            g: x,
-            b: c,
-        };
-    } else if (hsvPercent.h > 4 && hsvPercent.h <= 5) {
-        tempRgb = {
-            r: x,
-            g: 0,
-            b: c,
-        };
-    } else if (hsvPercent.h > 5 && hsvPercent.h <= 6) {
-        tempRgb = {
-            r: c,
-            g: 0,
-            b: x,
-        };
-    }
-
-    if (tempRgb) {
-        const m = hsvPercent.v - c;
-
-        result.r = Math.floor((tempRgb.r + m) * 255);
-        result.g = Math.floor((tempRgb.g + m) * 255);
-        result.b = Math.floor((tempRgb.b + m) * 255);
-    }
-
-    return result;
-}
-
-/**
- * Converts a Hue Saturation Value ([[IHsva]]) object in a [[IRgba]] object
- * @param hsva the Hue Saturation Value ([[IHsva]]) object
- * @returns the [[IRgba]] object
- */
-export function hsvaToRgba(hsva: IHsva): IRgba {
-    return {
-        a: hsva.a,
-        ...hsvToRgb(hsva),
-    };
-}
-
-/**
- * Converts a RGB ([[IRgb]]) object in a [[IHsv]] object
- * @param rgb the RGB ([[IRgb]]) object
- * @returns the [[IHsv]] object
- */
-export function rgbToHsv(rgb: IRgb): IHsv {
-    const rgbPercent = {
-            r: rgb.r / 255,
-            g: rgb.g / 255,
-            b: rgb.b / 255,
-        },
-        xMax = Math.max(rgbPercent.r, rgbPercent.g, rgbPercent.b),
-        xMin = Math.min(rgbPercent.r, rgbPercent.g, rgbPercent.b),
-        v = xMax,
-        c = xMax - xMin;
-
-    let h = 0;
-
-    if (v === rgbPercent.r) {
-        h = 60 * ((rgbPercent.g - rgbPercent.b) / c);
-    } else if (v === rgbPercent.g) {
-        h = 60 * (2 + (rgbPercent.b - rgbPercent.r) / c);
-    } else if (v === rgbPercent.b) {
-        h = 60 * (4 + (rgbPercent.r - rgbPercent.g) / c);
-    }
-
-    const s = !v ? 0 : c / v;
-
-    return {
-        h,
-        s: s * 100,
-        v: v * 100,
-    };
-}
-
-/**
- * Converts a RGB ([[IRgba]]) object in a [[IHsva]] object
- * @param rgba the RGB ([[IRgba]]) object
- */
-export function rgbaToHsva(rgba: IRgba): IHsva {
-    return {
-        a: rgba.a,
-        ...rgbToHsv(rgba),
-    };
-}
-
-/**
  * Returns a random ([[IRgb]]) color
  * @param min the minimum value for the color
  * @returns the random ([[IRgb]]) color
@@ -577,16 +318,6 @@ export function getStyleFromRgb(color: IRgb, opacity?: number): string {
  */
 export function getStyleFromHsl(color: IHsl, opacity?: number): string {
     return `hsla(${color.h}, ${color.s}%, ${color.l}%, ${opacity ?? 1})`;
-}
-
-/**
- * Gets a CSS style string from a [[IHsv]] object and opacity value
- * @param color the [[IHsv]] input color
- * @param opacity the opacity value
- * @returns the CSS style string
- */
-export function getStyleFromHsv(color: IHsv, opacity?: number): string {
-    return getStyleFromHsl(hsvToHsl(color), opacity);
 }
 
 export function colorMix(color1: IRgb | IHsl, color2: IRgb | IHsl, size1: number, size2: number): IRgb {