+++ engine/src/index.ts
@@ -1,6 +1,15 @@
 import { Engine } from "./engine";
+import { HslColorManager } from "./Utils/HslColorManager";
 import type { IOptions } from "./Options/Interfaces/IOptions";
 import type { RecursivePartial } from "./Types/RecursivePartial";
+import { RgbColorManager } from "./Utils/RgbColorManager";
+import { addColorManager } from "./Utils/ColorUtils";
+
+const rgbColorManager = new RgbColorManager(),
+    hslColorManager = new HslColorManager();
+
+addColorManager("rgb", rgbColorManager);
+addColorManager("hsl", hslColorManager);
 
 /**
  * The exposed tsParticles instance
@@ -10,10 +19,10 @@ const tsParticles = new Engine();
 tsParticles.init();
 
 export * from "./Core/Interfaces/Colors";
-export * from "./Core/Interfaces/Gradients";
 export * from "./Core/Interfaces/IBounds";
 export * from "./Core/Interfaces/IBubbleParticleData";
 export * from "./Core/Interfaces/ICircleBouncer";
+export * from "./Core/Interfaces/IColorManager";
 export * from "./Core/Interfaces/IContainerInteractivity";
 export * from "./Core/Interfaces/IContainerPlugin";
 export * from "./Core/Interfaces/ICoordinates";
@@ -73,12 +82,6 @@ export * from "./Enums/Types/EasingType";
 export * from "./Enums/AnimationStatus";
 export * from "./Enums/InteractivityDetect";
 export * from "./Options/Classes/AnimatableColor";
-export * from "./Options/Classes/Gradients/AnimatableGradient";
-export * from "./Options/Classes/Gradients/AnimatableGradientColor";
-export * from "./Options/Classes/Gradients/GradientAngle";
-export * from "./Options/Classes/Gradients/GradientAngleAnimation";
-export * from "./Options/Classes/Gradients/GradientColorOpacity";
-export * from "./Options/Classes/Gradients/GradientColorOpacityAnimation";
 export * from "./Options/Classes/AnimationOptions";
 export * from "./Options/Classes/Background/Background";
 export * from "./Options/Classes/BackgroundMask/BackgroundMask";
@@ -164,7 +167,6 @@ export * from "./Options/Interfaces/BackgroundMask/IBackgroundMaskCover";
 export * from "./Options/Interfaces/FullScreen/IFullScreen";
 export * from "./Options/Interfaces/IAnimatable";
 export * from "./Options/Interfaces/IAnimatableColor";
-export * from "./Options/Interfaces/IAnimatableGradient";
 export * from "./Options/Interfaces/IAnimation";
 export * from "./Options/Interfaces/IColorAnimation";
 export * from "./Options/Interfaces/IHslAnimation";
@@ -172,7 +174,6 @@ export * from "./Options/Interfaces/IManualParticle";
 export * from "./Options/Interfaces/IOptionLoader";
 export * from "./Options/Interfaces/IOptions";
 export * from "./Options/Interfaces/IOptionsColor";
-export * from "./Options/Interfaces/IOptionsGradient";
 export * from "./Options/Interfaces/IResponsive";
 export * from "./Options/Interfaces/IValueWithRandom";
 export * from "./Options/Interfaces/Interactivity/Events/IClickEvent";
@@ -250,8 +251,10 @@ export * from "./Types/SingleOrMultiple";
 export * from "./Types/PathOptions";
 export * from "./Utils/CanvasUtils";
 export * from "./Utils/ColorUtils";
+export * from "./Utils/HslColorManager";
 export * from "./Utils/NumberUtils";
 export * from "./Utils/OptionsUtils";
+export * from "./Utils/RgbColorManager";
 export * from "./Utils/Utils";
 export { tsParticles };