+++ plugins/hsvColor/src/index.ts
@@ -1,0 +1,6 @@
+import { HsvColorManager } from "./HsvColorManager";
+import { addColorManager } from "tsparticles-engine";
+
+export function loadHsvColorPlugin(): void {
+    addColorManager("hsv", new HsvColorManager());
+}