+++ utils/webpack/plugins/buildPlugin.js
@@ -1,7 +1,7 @@
-const {getPluginEntry} = require("./getPluginEntry");
-const {getConfig} = require("../common/getConfig");
+const { getPluginEntry } = require("./getPluginEntry");
+const { getConfig } = require("../common/getConfig");
 
-function loadParticlesPlugin(moduleName, pluginName, version, dir) {
+function loadParticlesPlugin(moduleName, pluginName, version, dir, bundle) {
     const banner = `Author : Matteo Bruni
 MIT license: https://opensource.org/licenses/MIT
 Demo / Generator : https://particles.js.org/
@@ -9,9 +9,12 @@ GitHub : https://www.github.com/matteobruni/tsparticles
 How to use? : Check the GitHub README
 v${version}`, minBanner = `tsParticles ${pluginName} Plugin v${version} by Matteo Bruni`;
 
-    return [
+    return bundle ? [
+        getConfig(getPluginEntry(moduleName, false), banner, minBanner, dir, false),
+        getConfig(getPluginEntry(moduleName, true), banner, minBanner, dir, true)
+    ] : [
         getConfig(getPluginEntry(moduleName, false), banner, minBanner, dir, false)
     ];
 }
 
-module.exports = {loadParticlesPlugin};
+module.exports = { loadParticlesPlugin };