+++ plugins/infection/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/matteobruni/tsparticles/compare/tsparticles-plugin-infection@2.1.3...tsparticles-plugin-infection@2.1.4) (2022-07-28)
+
+**Note:** Version bump only for package tsparticles-plugin-infection
+
+
+
+
+
 ## [2.1.3](https://github.com/matteobruni/tsparticles/compare/tsparticles-plugin-infection@2.1.2...tsparticles-plugin-infection@2.1.3) (2022-07-01)
 
 **Note:** Version bump only for package tsparticles-plugin-infection